+++ admin/remoter.go
@@ -21,6 +21,14 @@ type RemoteOptions struct {
 	LogEmitter   app.LogEmitter
 }
 
+type RunOptions struct {
+	// Container represents the specific container name to execute against in the k8s pod/ecs task
+	Container   string
+	Username    string
+	LogStreamer app.LogStreamer
+	LogEmitter  app.LogEmitter
+}
+
 type Remoter interface {
 	// Exec allows a user to execute a command (usually tunneling) into a running service
 	// This only makes sense for container-based providers
@@ -28,4 +36,7 @@ type Remoter interface {
 
 	// Ssh allows a user to SSH into a running service
 	Ssh(ctx context.Context, options RemoteOptions) error
+
+	// Run starts a new job/task and executes a command
+	Run(ctx context.Context, options RunOptions, cmd []string) error
 }